[#61] - New structure implemeted
