+++ assets/js/constants/ClickDisablerConstants.js
@@ -1,0 +1,9 @@
+/**
+ * Constants used for ClickDisabler
+ */
+export class ClickDisablerConstants {
+  /**
+       * Selector for disabled elements in the DOM.
+       */
+  static disabledElementSelector = '.disabled';
+}
