+++ assets/js/shortcodes/index.js
@@ -1,0 +1,3 @@
+export { CollapsibleBlock } from './CollapsibleBlock';
+export { CollapsibleGroupCommand } from './CollapsibleGroupCommand';
+export { ClickDisabler } from './ClickDisabler';
